backend/requirements.txt
# Kronaghor Backend Requirements

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlmodel>=0.0.14

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI Providers
openai>=1.10.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
